Create BaseTest.java. Refactoring  code
